requirements.txt
Flask==3.0.2
google-genai
python-dotenv==1.0.1